Refactor hierarchy
